cli: refactor get commands into a module
